feat: init crud

+++ src/app.service.ts
@@ -5,10 +5,14 @@ import { UserRepository } from './repositories/user-repository';
 import { ReadUserRequestBody } from './dtos/read-user-request-body';
 import { UpdateUserRequestBody } from './dtos/update-user-request-body';
 import { DeleteUserRequestBody } from './dtos/delete-user-request-body';
+import { HashProvider } from './provider/hash-provider';
 
 @Injectable()
 export class AppService {
-  constructor(private userRepository: UserRepository) {}
+  constructor(
+    private userRepository: UserRepository,
+    private hashProvider: HashProvider,
+  ) {}
   async create({
     name,
     email,
@@ -18,7 +22,8 @@ export class AppService {
     if (emailInUse) {
       throw new Error('');
     }
-    const parcialUser = new User({ name, email, password });
+    const hashPassword = this.hashProvider.hashPassword(password);
+    const parcialUser = new User({ name, email, password: hashPassword });
     const defineUser = await this.userRepository.create(parcialUser);
     return defineUser;
   }
@@ -30,13 +35,19 @@ export class AppService {
     const user = await this.userRepository.read(id);
     return user;
   }
+  
   async update({ id, name }: UpdateUserRequestBody) {
     const parcialUser = await this.userRepository.read(id);
     parcialUser.name = name;
     const defineUser = await this.userRepository.update(parcialUser);
     return defineUser;
   }
+
   async delete({ id }: DeleteUserRequestBody) {
+    const userExit = await this.userRepository.read(id);
+    if (userExit === null) {
+      throw new Error('mama mia');
+    }
     await this.userRepository.delete(id);
   }
 }
